requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
Flask>=2.0.0
flask-cors>=3.0.10
gunicorn>=20.1.0
kaggle>=1.5.12
pandas 